backend_py/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
